requirements.txt
anyio==3.7.1
certifi==2023.5.7
cffi==1.15.1
click==8.1.4
colorama==0.4.6
cryptography==41.0.1
ecdsa==0.18.0
fastapi==0.99.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.11
python-dotenv
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
